👌 IMPROVE: Improve timetable course item
