+++ src/components/Timetable/index.jsx
@@ -1,4 +1,3 @@
 export { default as TimetableContainer } from './TimetableContainer'
 export { default as TimetableCourseItem } from './TimetableCourseItem'
-export { default as TimetableCourseLectureItem } from './TimetableCourseLectureItem'
 export { default as TimetableLayout } from './TimetableLayout'
